requirements.txt
python-telegram-bot==13.4.1
PyDrive2==1.10.0
requests==2.25.1
beautifulsoup4==4.9.3
python-gitlab==2.6.0
matplotlib==3.3.4
pandas==1.2.3
Pillow==8.1.2
lxml==4.6.3
